# do not reallocate memory when processing shared channels