--- a/glommio/Cargo.toml
+++ b/glommio/Cargo.toml
@@ -42,7 +42,7 @@ crossbeam = "0.8"
 [dev-dependencies]
 futures = "0.3.5"
 fastrand = "1.4.0"
-tokio = { version = "0.3.5", default-features = false, features = ["rt", "macros", "rt-multi-thread", "net", "io-util", "time"] }
+tokio = { version = "0.3.5", default-features = false, features = ["rt", "macros", "rt-multi-thread", "net", "io-util", "time", "sync"] }
 rand = "0.8.0"
 tracing-subscriber = "0.2"
 pretty_env_logger = "0.4"