### What does this PR do?

The implementation used to iterate over the tree of wakers and would
recreate a new tree with the remaining ones. This is not efficient as it
leads to a high number of allocations. Instead, iterate over the
existing tree and reuse the memory in place.

Additionally:
* use the same tree to store wakers and checker functions to reduce
the number of lookups and simplify the lifecycle of both;
* use a `SmallVec` to store wakers and optimize for the case
where a single waker is consuming the shared channel.

I added some additional benchmarks to exercise this code path. This
is a small but measurable improvement, especially in the many-to-many
benchmark (which is really meant to be worst-case). See below.

Before:
```
cost of mesh message shared channel: 21.48µs, peers 8, capacity 1024
cost of mesh message shared channel: 32.648µs, peers 8, capacity 1
```

After:
```
cost of mesh message shared channel: 20.744µs, peers 8, capacity 1024
cost of mesh message shared channel: 25.872µs, peers 8, capacity 1
```

### Motivation

What inspired you to submit this pull request?

### Related issues

A list of issues either fixed, containing architectural discussions, otherwise relevant
for this Pull Request.

### Additional Notes

Anything else we should know when reviewing?

### Checklist

[] I have added unit tests to the code I am submitting
[] My unit tests cover both failure and success scenarios
[] If applicable, I have discussed my architecture